+++ pytket/extensions/qiskit/backends/ibm.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import os
 import itertools
 import logging
 from ast import literal_eval
@@ -125,8 +124,6 @@ def _save_ibmq_auth(qiskit_config: Optional[QiskitConfig]) -> None:
     token = None
     if qiskit_config is not None:
         token = qiskit_config.ibmq_api_token
-    if token is None and os.getenv("PYTKET_REMOTE_QISKIT_TOKEN") is not None:
-        token = os.getenv("PYTKET_REMOTE_QISKIT_TOKEN")
     try:
         if token is not None:
             IBMProvider.save_account(token, overwrite=True)