+++ docs/changelog.rst
@@ -1,6 +1,14 @@
 Changelog
 ~~~~~~~~~
 
+0.41.0 (July 2023)
+------------------
+
+* Update pytket version requirement to 1.17.
+* Fix conversion of qiskit `UnitaryGate` to and from pytket (up to 3 qubits).
+* Fix handling of qiskit controlled gates in the :py:meth:`qiskit_to_tk` converter.
+* Handle CCZ and CSX gates in circuit converters.
+
 0.40.0 (June 2023)
 ------------------